Refactor: OutputView클래스 view패키지로 이동
