+++ src/main/java/racingcar/view/OutputView.java
@@ -1,4 +1,4 @@
-package racingcar.utils;
+package racingcar.view;
 
 import racingcar.domain.Car;
